refactor: avoid code duplication in tasks
