+++ tasks/removeTokens.ts
@@ -1,6 +1,6 @@
 import { ContractTransaction } from 'ethers';
 import { HardhatRuntimeEnvironment } from 'hardhat/types';
-import { getExistingConfig } from './deploy';
+import { getVerifiers } from './utils';
 
 export const removeTokens = async (
   taskArgs: { tokenlist: string },
@@ -8,96 +8,14 @@ export const removeTokens = async (
 ) => {
   const tokenAddresses = taskArgs.tokenlist.split(',');
 
-  const { ethers, network } = hre;
-
-  if (!network) {
-    throw new Error('Unknown Network');
-  }
-
-  const { chainId } = network.config;
-
-  if (!chainId) {
-    throw new Error('Unknown Chain Id');
-  }
-
-  const contractAddresses = getExistingConfig();
-
-  if (!contractAddresses) {
-    throw new Error('No contracts deployed');
-  }
-
-  const networkChainKey = `${network.name}.${chainId}`;
-  const contractAddressesDeployed = contractAddresses[networkChainKey];
-
-  if (!contractAddressesDeployed) {
-    throw new Error(`Contracts not deployed for chain ID ${chainId}`);
-  }
-
-  const deployVerifierAddress = contractAddressesDeployed.DeployVerifier;
-  const relayVerifierAddress = contractAddressesDeployed.RelayVerifier;
-  const customDeployVerifierAddress =
-    contractAddressesDeployed.CustomSmartWalletDeployVerifier;
-  const customRelayVerifierAddress =
-    contractAddressesDeployed.CustomSmartWalletRelayVerifier;
-  const nativeDeployVerifierAddress =
-    contractAddressesDeployed.NativeHolderSmartWalletDeployVerifier;
-  const nativeRelayVerifierAddress =
-    contractAddressesDeployed.NativeHolderSmartWalletDeployVerifier;
-
-  // TODO: This need to be refactored
-  if (!deployVerifierAddress) {
-    throw new Error('Could not obtain deploy verifier address');
-  }
-
-  if (!relayVerifierAddress) {
-    throw new Error('Could not obtain relay verifier address');
-  }
-
-  if (!customDeployVerifierAddress) {
-    throw new Error('Could not obtain custom deploy verifier address');
-  }
-
-  if (!customRelayVerifierAddress) {
-    throw new Error('Could not obtain custom deploy verifier address');
-  }
-
-  if (!nativeDeployVerifierAddress) {
-    throw new Error(
-      'Could not obtain deploy verifier address for the NativeHolderSmartWallet'
-    );
-  }
-
-  if (!nativeRelayVerifierAddress) {
-    throw new Error(
-      'Could not obtain relay verifier address for the NativeHolderSmartWallet'
-    );
-  }
-
-  const deployVerifier = await ethers.getContractAt(
-    'DeployVerifier',
-    deployVerifierAddress
-  );
-  const relayVerifier = await ethers.getContractAt(
-    'RelayVerifier',
-    relayVerifierAddress
-  );
-  const customDeployVerifier = await ethers.getContractAt(
-    'CustomSmartWalletDeployVerifier',
-    customDeployVerifierAddress
-  );
-  const customRelayVerifier = await ethers.getContractAt(
-    'RelayVerifier',
-    customRelayVerifierAddress
-  );
-
-  const nativeHolderDeployVerifier = await ethers.getContractAt(
-    'DeployVerifier',
-    nativeDeployVerifierAddress
-  );
-  const nativeHolderRelayVerifier = await ethers.getContractAt(
-    'RelayVerifier',
-    nativeRelayVerifierAddress
-  );
+  const {
+    deployVerifier,
+    relayVerifier,
+    customDeployVerifier,
+    customRelayVerifier,
+    nativeHolderDeployVerifier,
+    nativeHolderRelayVerifier,
+  } = await getVerifiers(hre);
 
   const verifierMap: Map<
     string,
