+++ test/tasks/removeTokens.test.ts
@@ -6,6 +6,7 @@ import * as hre from 'hardhat';
 import { ethers } from 'hardhat';
 import sinon from 'sinon';
 import { removeTokens } from '../../tasks/removeTokens';
+import { stubReadFileSync } from './utils';
 
 use(chaiAsPromised);
 
@@ -15,32 +16,9 @@ describe('Remove Tokens Script', function () {
       tokenlist: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
     };
 
-    const contractAddresses = {
-      Penalizer: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      RelayHub: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      SmartWallet: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      SmartWalletFactory: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      DeployVerifier: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      RelayVerifier: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      CustomSmartWallet: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      CustomSmartWalletFactory: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      CustomSmartWalletDeployVerifier:
-        '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      CustomSmartWalletRelayVerifier:
-        '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      VersionRegistry: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-      UtilToken: '0x145845fd06c85B7EA1AA2d030E1a747B3d8d15D7',
-    };
-
-    const chainContractAddresses = {
-      'hardhat.33': contractAddresses,
-    };
-
     beforeEach(function () {
       sinon.stub(fs, 'existsSync').returns(true);
-      sinon
-        .stub(fs, 'readFileSync')
-        .returns(JSON.stringify(chainContractAddresses));
+      stubReadFileSync();
       hre.network.config.chainId = 33;
     });
